requirements.txt
fastapi
uvicorn
sqlalchemy
databases
pydantic
python-dotenv
mysql-connector-python
